feat(config): dedicated configs for Payload CMS
